app stopp script survvives app not running

+++ hooks/appStop.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 set -e
 
-systemctl stop bidenbot-discord
-systemctl stop bidenbot-slack
+if systemctl is-active bidenbot-discord; then
+  systemctl stop bidenbot-discord
+fi
+if systemctl is-active bidenbot-slack; then
+  systemctl stop bidenbot-slack
+fi
